+++ services/frontend/src/app/core/interfaces/digital-service.interfaces.ts
@@ -419,7 +419,7 @@ export interface RenewServiceUpdateResp {
 export interface GraphDescriptionContent {
     description: string;
     scale: string;
-    textDescription: string;
+    textDescription?: string;
     analysis?: string;
     toGoFurther?: string;
     textResourceDescription?: string;